moved release-notes related coding to github/release_notes.py removed dead code plus some minor refactorings
