+++ cli/githubutil.py
@@ -15,11 +15,6 @@
 import version
 from urllib.parse import urlparse, parse_qs
 from github3.exceptions import NotFoundError
-from pydash import _
-from collections import namedtuple
-from semver import parse_version_info
-import git
-import re
 
 from util import ctx, not_empty, info, warning, fail, verbose, CliHint, CliHints, existing_dir
 from github.webhook import GithubWebHookSyncer, WebhookQueryAttributes
@@ -30,17 +25,10 @@ from github.util import (
     _add_user_to_team,
     _add_all_repos_to_team
 )
+from github.release_notes import generate_release_notes
 import product.model
 
-ReleaseNote = namedtuple('ReleaseNote', ["category_id", "group_id", "text", "pr_number", "user_login"])
-Category = namedtuple("Category", "identifier title")
-Group = namedtuple("Group", "identifier title")
-categories = \
-    Category(identifier='noteworthy', title='## Most notable changes'),\
-    Category(identifier='improvement', title='## Improvements')
-groups = \
-    Group(identifier='user', title='### To end users'), \
-    Group(identifier='operator', title='### To operations team')
+
 
 def assign_github_team_to_repo(
     github_cfg_name: str,
@@ -102,197 +90,6 @@ def generate_release_notes_cli(
     )
     generate_release_notes(repo_dir=repo_dir, helper=helper, repository_branch=repository_branch, commit_range=commit_range)
 
-def generate_release_notes(
-    repo_dir: str,
-    helper: GitHubRepositoryHelper,
-    repository_branch: str,
-    commit_range: str=None
-):
-    repo = git.Repo(existing_dir(repo_dir))
-    if repo.active_branch.name != repository_branch:
-        fail('need to switch to branch {branch}. Current active branch: {active_branch}'.format(branch=repository_branch, active_branch=repo.active_branch.name))
-
-    if not commit_range:
-        commit_range = calculate_range(repo, helper)
-    pr_numbers = fetch_pr_numbers_in_range(repo, commit_range)
-    release_note_objs = fetch_release_notes_from_prs(helper, pr_numbers)
-    release_notes = build_markdown(release_note_objs)
-
-    info(release_notes)
-    return release_notes
-
-def build_markdown(
-    release_note_objs: list
-) -> str:
-    def release_notes_for_group(
-        category: Category,
-        group: Group,
-        release_note_objs: list()
-    ) -> list:
-        release_note_objs = _.filter(release_note_objs, lambda release_note: category.identifier == release_note.category_id and group.identifier == release_note.group_id)
-
-        release_note_lines = list()
-        if release_note_objs:
-            release_note_lines.append(group.title)
-            for release_note in release_note_objs:
-                for i, rls_note_line in enumerate(release_note.text.splitlines()):
-                    if i == 0:
-                        release_note_lines.append('* {rls_note_line} (#{pr_num}, @{user})'.format(pr_num=release_note.pr_number, user=release_note.user_login, rls_note_line=rls_note_line))
-                    else:
-                        release_note_lines.append('  * {rls_note_line}'.format(rls_note_line=rls_note_line))
-        return release_note_lines
-
-    def release_notes_for_category(
-        category: Category,
-        release_note_objs: list
-    ) -> list:
-        rn_lines = list()
-        rn_lines_category = list()
-        for group in groups:
-            rn_lines_category.extend(release_notes_for_group(category=category, group=group, release_note_objs=release_note_objs))
-
-        if rn_lines_category:
-            rn_lines.append(category.title)
-            rn_lines.extend(rn_lines_category)
-        return rn_lines
-
-    release_note_lines = list()
-    for category in categories:
-        release_note_lines.extend(release_notes_for_category(category, release_note_objs=release_note_objs))
-
-    if release_note_lines:
-        return '\n'.join(release_note_lines)
-    else: # fallback
-        return 'release notes'
-
-def calculate_range(
-    repo: git.Repo,
-    helper: GitHubRepositoryHelper,
-) -> str:
-
-    branch_head = repo.head.commit
-    range_start = _.head(reachable_release_tags_from_commit(helper, repo, branch_head))
-
-    range_end = None
-    try:
-        range_end = repo.git.describe(branch_head) # better readable range_end by describing head commit
-    except git.exc.GitCommandError:
-        range_end = branch_head.hexsha
-
-    commit_range = "{start}..{end}".format(start=range_start, end=range_end)
-    return commit_range
-
-def release_tags(
-    helper: GitHubRepositoryHelper,
-    repo: git.Repo
-) -> list:
-    release_tags = helper.release_tags()
-    tags = _ \
-        .chain(repo.tags) \
-        .map(lambda tag: {"tag": tag.name, "commit": tag.commit.hexsha}) \
-        .filter(lambda item: _.find(release_tags, lambda el: el == item['tag'])) \
-        .key_by('commit') \
-        .map_values('tag') \
-        .value()
-    return tags
-
-def reachable_release_tags_from_commit(
-    helper: GitHubRepositoryHelper,
-    repo: git.Repo, commit
-) -> list:
-    tags = release_tags(helper, repo)
-
-    visited = set()
-    queue = list()
-    queue.append(commit)
-    visited.add(commit.hexsha)
-
-    reachable_tags = list()
-
-    while queue:
-        commit = queue.pop(0)
-        if commit.hexsha in tags:
-            reachable_tags.append(tags[commit.hexsha])
-        not_visited_parents = _.filter(commit.parents, lambda parent_commit: not parent_commit.hexsha in visited)
-        if not_visited_parents:
-            queue.extend(not_visited_parents)
-            # queue.sort(key=lambda commit: commit.committed_date, reverse=True) #not needed anymore as we sort by semver at the end
-            visited |= set(_.map(not_visited_parents, lambda commit: commit.hexsha))
-
-    reachable_tags.sort(key=lambda t: parse_version_info(t), reverse=True)
-
-    if not reachable_tags:
-        warning('no release tag found, falling back to root commit')
-        root_commits = repo.iter_commits(rev=commit, max_parents=0)
-        root_commit = next(root_commits, None)
-        if not root_commit:
-            fail('could not determine root commit from rev {rev}'.format(rev=commit.hexsha))
-        if next(root_commits, None):
-            fail('cannot determine range for release notes. Repository has multiple root commits. Specify range via commit_range parameter.')
-        reachable_tags.append(root_commit.hexsha)
-
-    return reachable_tags
-
-
-def fetch_pr_numbers_in_range(
-    repo: git.Repo,
-    commit_range: str
-) -> set:
-    info('git log {range}'.format(range=commit_range))
-    gitLogs = repo.git.log(commit_range, pretty='%s').splitlines()
-    pr_numbers = []
-    for commitMessage in gitLogs:
-        if commitMessage.startswith('Merge pull'):
-            pr_number = _.head(re.findall(r"#(\d+|$)", commitMessage))
-            if pr_number:
-                pr_numbers.append(pr_number)
-
-    verbose('Merged pull request numbers in range {range}: {pr_numbers}'.format(range=commit_range, pr_numbers=pr_numbers))
-    return pr_numbers
-
-def fetch_release_notes_from_prs(
-    helper: GitHubRepositoryHelper,
-    pr_numbers_in_range: set
-) -> list:
-    # we should consider adding a release-note label to the PRs to reduce the number of search results
-    prs_iter = helper.search_issues_in_repo('type:pull is:closed')
-
-    release_notes = list()
-    for pr_iter in prs_iter:
-        pr_dict = pr_iter.as_dict()
-
-        pr_number = pr_dict['number']
-        if not str(pr_number) in pr_numbers_in_range:
-            continue
-
-        release_notes_pr = extract_release_notes(pr_number=pr_number, text=pr_dict['body'], user_login=_.get(pr_dict, 'user.login'))
-        if not release_notes_pr:
-            continue
-
-        release_notes.extend(release_notes_pr)
-    return release_notes
-
-def extract_release_notes(
-    pr_number: int,
-    text: str,
-    user_login: str
-) -> list:
-    release_notes = list()
-
-    quotes = re.findall(r"``` *(improvement|noteworthy)( (user|operator)?)?.*?\n(.*?)\n```", text, re.MULTILINE | re.DOTALL)
-    for quote in quotes:
-        quote = _.map(quote, lambda obj: _.trim(obj))
-
-        text = quote[3]
-        if not text or 'none' == text.lower():
-            continue
-
-        category = quote[0]
-        group = quote[2] or 'user'
-
-        release_notes.append(ReleaseNote(category_id=category, group_id=group, text=text, pr_number=pr_number, user_login=user_login))
-    return release_notes
-
 def release_and_prepare_next_dev_cycle(
     github_cfg_name: str,
     github_repository_owner: str,
@@ -300,7 +97,7 @@ def release_and_prepare_next_dev_cycle(
     repository_branch: str,
     repository_version_file_path: str,
     release_version: str,
-    release_notes: str=None,
+    generate_release_notes: bool,
     version_operation: str="bump_minor",
     prerelease_suffix: str="dev",
     author_name: str="gardener-ci",
@@ -324,8 +121,10 @@ def release_and_prepare_next_dev_cycle(
             )
         )
 
-    if not release_notes:
+    if generate_release_notes:
         release_notes = generate_release_notes(repo_dir=repo_dir, helper=helper, repository_branch=repository_branch, commit_range=commit_range)
+    else:
+        release_notes = 'release notes'
 
     # Do all the version handling upfront to catch errors early
     # Bump release version and add suffix
